refactor(core): extract app state and init logic
This commit begins the modularization of the application by extracting core logic from the monolithic app.js file.

- A new `core/` directory has been created to house foundational code.
- The global `app` state object has been moved into its own module, `core/state.js`, to centralize state management.
- The main application logic, including the `init` function, has been moved to `core/main.js`.
- The root `app.js` is now a simple entry point that imports and calls `init()`.

This change improves separation of concerns without altering functionality, paving the way for further refactoring.
